examples/langchain/langgraph_agent/pyproject.toml
[project]
name = "langgraph-agent"
version = "0.1.0"
description = "LangGraph (12-factor) agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.27",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.2",
    "python-dotenv>=1.1.1",
    "redis==5.3.1",
    "langgraph-checkpoint-redis>=0.0.8",
    "nanoid>=2.0.0",
    "langgraph-cli[inmem]>=0.3.8",
]
